Simple sphere model for the dependencies
