--- app/components/pokemon-visualizer/pokemon.js
@@ -1,20 +1,0 @@
-import React, { useRef }  from 'react'
-import { useLoader } from 'react-three-fiber'
-import { OBJLoader } from 'three/examples/jsm/loaders/OBJLoader'
-
-const Pokemon = props => {
-  const ref = useRef()
-//   useFrame(() => (ref.current.rotation.x = ref.current.rotation.y += 0.01))
-  return (
-    <mesh
-      ref={ref}
-      onClick={e => console.log('click')}
-      onPointerOver={e => console.log('hover')}
-      onPointerOut={e => console.log('unhover')}>
-      <boxBufferGeometry attach="geometry" args={[1, 1, 1]} />
-      <meshNormalMaterial attach="material" />
-    </mesh>
-  )
-}
-
-export default Scene
