+++ app/components/threejs-visualizer/scene.js
